+++ Coimbra/AssemblyInfo.cs
@@ -1,5 +1,7 @@
 using System.Runtime.CompilerServices;
+using UnityEngine.Scripting;
 
+[assembly: AlwaysLinkAssembly]
 [assembly: InternalsVisibleTo("Coimbra.Editor")]
 [assembly: InternalsVisibleTo("Coimbra.Editor.Tests")]
 [assembly: InternalsVisibleTo("Coimbra.ExtendedInspector")]