+++ Coimbra/Events/ApplicationQuitEvent.cs
@@ -2,6 +2,9 @@
 
 namespace Coimbra
 {
+    /// <summary>
+    /// Invoked during <a href="https://docs.unity3d.com/ScriptReference/MonoBehaviour.OnApplicationQuit.html">OnApplicationQuit</a>.
+    /// </summary>
     [Preserve]
     public readonly struct ApplicationQuitEvent { }
 }