+++ Coimbra/Systems/EventSystem.cs
@@ -2,12 +2,9 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Coimbra.Systems
+namespace Coimbra
 {
-    /// <summary>
-    /// Default implementation for <see cref="IEventService"/>.
-    /// </summary>
-    public class EventSystem : IEventService
+    internal sealed class EventSystem : IEventService
     {
         private sealed class Event
         {
@@ -35,15 +32,15 @@ namespace Coimbra.Systems
         private readonly object _serviceKey;
         private readonly Dictionary<Type, Event> _events = new Dictionary<Type, Event>();
 
-        public EventSystem(object serviceKey = null)
+        private EventSystem(object serviceKey = null)
         {
             _serviceKey = serviceKey;
         }
 
-        /// <inheritdoc cref="IService.OwningLocator"/>
+        /// <inheritdoc/>
         public ServiceLocator OwningLocator { get; set; }
 
-        /// <inheritdoc cref="IEventService.AddListener{T}(EventHandler{T})"/>.
+        /// <inheritdoc/>
         public EventHandle AddListener<T>(EventHandler<T> eventCallback)
         {
             if (eventCallback == null)
@@ -57,7 +54,7 @@ namespace Coimbra.Systems
             });
         }
 
-        /// <inheritdoc cref="IEventService.AddListener{T}(EventRefHandler{T})"/>.
+        /// <inheritdoc/>
         public EventHandle AddListener<T>(EventRefHandler<T> eventCallback)
         {
             if (eventCallback == null)
@@ -79,56 +76,56 @@ namespace Coimbra.Systems
             return handle;
         }
 
-        /// <inheritdoc cref="IEventService.CompareEventKey{T}"/>.
+        /// <inheritdoc/>
         public bool CompareEventKey<T>(object eventKey)
         {
             return CompareEventKey(typeof(T), eventKey);
         }
 
-        /// <inheritdoc cref="IEventService.CompareEventKey"/>.
+        /// <inheritdoc/>
         public bool CompareEventKey(Type eventType, object eventKey)
         {
             return _events.TryGetValue(eventType, out Event e) ? e.Key == eventKey : eventKey == null;
         }
 
-        /// <inheritdoc cref="IDisposable.Dispose"/>
+        /// <inheritdoc/>
         public void Dispose()
         {
             RemoveAllListeners(_serviceKey);
             _events.Clear();
         }
 
-        /// <inheritdoc cref="IEventService.HasAnyListeners{T}"/>.
+        /// <inheritdoc/>
         public bool HasAnyListeners<T>()
         {
             return HasAnyListeners(typeof(T));
         }
 
-        /// <inheritdoc cref="IEventService.HasAnyListeners"/>.
+        /// <inheritdoc/>
         public bool HasAnyListeners(Type eventType)
         {
             return _events.TryGetValue(eventType, out Event e) && e.Handles.Count > 0;
         }
 
-        /// <inheritdoc cref="IEventService.HasListener"/>.
+        /// <inheritdoc/>
         public bool HasListener(in EventHandle eventHandle)
         {
             return eventHandle.IsValid && _events.TryGetValue(eventHandle.Type, out Event e) && e.Handles.Contains(eventHandle);
         }
 
-        /// <inheritdoc cref="IEventService.Invoke"/>.
+        /// <inheritdoc/>
         public void Invoke(object eventSender, Type eventType, object eventKey = null, bool ignoreException = false)
         {
             Invoke(eventSender, Activator.CreateInstance(eventType), eventKey, ignoreException);
         }
 
-        /// <inheritdoc cref="IEventService.Invoke{T}(object, T, object, bool)"/>.
+        /// <inheritdoc/>
         public void Invoke<T>(object eventSender, T eventData, object eventKey = null, bool ignoreException = false)
         {
             Invoke(eventSender, ref eventData, eventKey, ignoreException);
         }
 
-        /// <inheritdoc cref="IEventService.Invoke{T}(object, ref T, object, bool)"/>.
+        /// <inheritdoc/>
         public void Invoke<T>(object eventSender, ref T eventData, object eventKey = null, bool ignoreException = false)
         {
             if (!_events.TryGetValue(typeof(T), out Event e))
@@ -152,7 +149,7 @@ namespace Coimbra.Systems
             }
         }
 
-        /// <inheritdoc cref="IEventService.RemoveAllListeners(object, bool)"/>.
+        /// <inheritdoc/>
         public void RemoveAllListeners(object serviceKey = null, bool ignoreException = false)
         {
             if (_serviceKey != null && _serviceKey != serviceKey)
@@ -176,13 +173,13 @@ namespace Coimbra.Systems
             }
         }
 
-        /// <inheritdoc cref="IEventService.RemoveAllListeners{T}"/>.
+        /// <inheritdoc/>
         public void RemoveAllListeners<T>(object eventKey = null, bool ignoreException = false)
         {
             RemoveAllListeners(typeof(T), eventKey, ignoreException);
         }
 
-        /// <inheritdoc cref="IEventService.RemoveAllListeners(Type, object, bool)"/>.
+        /// <inheritdoc/>
         public void RemoveAllListeners(Type eventType, object eventKey = null, bool ignoreException = false)
         {
             if (!_events.TryGetValue(eventType, out Event e))
@@ -208,7 +205,7 @@ namespace Coimbra.Systems
             e.Handles.Clear();
         }
 
-        /// <inheritdoc cref="IEventService.RemoveListener"/>.
+        /// <inheritdoc/>
         public void RemoveListener(in EventHandle eventHandle)
         {
             if (!_events.TryGetValue(eventHandle.Type, out Event e))
@@ -222,13 +219,13 @@ namespace Coimbra.Systems
             }
         }
 
-        /// <inheritdoc cref="IEventService.ResetEventKey{T}"/>.
+        /// <inheritdoc/>
         public void ResetEventKey<T>(object eventKey)
         {
             ResetEventKey(typeof(T), eventKey);
         }
 
-        /// <inheritdoc cref="IEventService.ResetEventKey"/>.
+        /// <inheritdoc/>
         public void ResetEventKey(Type eventType, object eventKey)
         {
             if (eventKey == null || !_events.TryGetValue(eventType, out Event e))
@@ -244,7 +241,7 @@ namespace Coimbra.Systems
             e.Key = null;
         }
 
-        /// <inheritdoc cref="IEventService.SetEventKey{T}"/>.
+        /// <inheritdoc/>
         public void SetEventKey<T>(object eventKey)
         {
             if (_events.TryGetValue(typeof(T), out Event e))